+++ cmd/netductor/main.go
@@ -10,7 +10,7 @@ import (
 	"github.com/PavelNeyman/netductor/internal/ndconfig"
 )
 
-var version = "0.8.4"
+var version = "0.8.5"
 
 func main() {
 	ndconfig.Load()